Refactored controller structure
