+++ cut-be/src/app/controllers/v1/mod.rs
@@ -1,3 +1,2 @@
 pub mod cut;
 pub mod ping;
-pub mod snippet;
