+++ cut-be/src/app/datatransfers/cut.rs
@@ -5,7 +5,6 @@ pub struct Cut {
     pub name: String,
     #[serde(skip)]
     pub owner: String,
-    #[serde(skip_deserializing)]
     pub variant: String,
     pub metadata: String,
     pub data: String,
