+++ cut-be/src/app/handlers/cut.rs
@@ -1,8 +1,12 @@
-use crate::app::{datatransfers::cut::Cut, Module};
+use crate::app::{constants::HASH_LENGTH, datatransfers::cut::Cut, Module};
 use crate::core::error::{HandlerError, HandlerErrorKind};
+use crate::utils::hash;
 use actix_web::web;
 use r2d2_redis::redis::Commands;
-use std::collections::HashMap;
+use std::{
+    collections::HashMap,
+    time::{SystemTime, UNIX_EPOCH},
+};
 
 pub fn get_one(m: web::Data<Module>, id: String) -> Result<Cut, HandlerError> {
     let rd = &mut m.rd_pool.get()?;
@@ -38,6 +42,33 @@ pub fn get_one(m: web::Data<Module>, id: String) -> Result<Cut, HandlerError> {
                     .parse()?,
             })
         }
+        Err(e) => Err(e.into()),
+    }
+}
+
+pub fn insert(
+    m: web::Data<Module>,
+    user_subject: String,
+    cut: Cut,
+) -> Result<String, HandlerError> {
+    let rd = &mut m.rd_pool.get()?;
+    let hash: String = hash::generate(HASH_LENGTH).into();
+    let created_at = match SystemTime::now().duration_since(UNIX_EPOCH) {
+        Ok(duration) => duration.as_secs(),
+        Err(_) => return Err(HandlerErrorKind::GeneralError.into()),
+    };
+    match rd.hset_multiple::<String, &str, String, String>(
+        hash.clone(),
+        &[
+            ("name", cut.name),
+            ("owner", user_subject),
+            ("variant", cut.variant),
+            ("metadata", cut.metadata),
+            ("data", cut.data),
+            ("created_at", created_at.to_string()),
+        ],
+    ) {
+        Ok(_) => Ok(hash),
         Err(e) => Err(e.into()),
     }
 }
