+++ cut-be/src/app/handlers/mod.rs
@@ -1,3 +1,2 @@
 pub mod auth;
 pub mod cut;
-pub mod snippet;
